{{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/users/api/serializers.py
import logging
from rest_framework import serializers
from {{ cookiecutter.project_slug }}.users.models import User

logger = logging.getLogger(__name__)


class UserSerializer(serializers.ModelSerializer):

    class Meta:
        model = User
        exclude = ('password',)


class UserDetailSerializer(serializers.ModelSerializer):

    class Meta:
        model = User
        fields = ('name',)